Update patch set 2
Patch Set 2:

> actually nougat-release isn't the right branch for CTS release, pls
 > cp to nougat-cts-dev branch.

nougat-cts-dev is a manifest branch that actually points to nougat-release . I don't think there's a way to cp to nougat-cts-dev, it's not even an option when I click the "cherry-pick" button. Well, actually there's a to cp there, that is cp'ing to nougat-release :)

I tried from the command line as well, and it just created another patch for this change, so it really looks there's no way to cherrypick there.

Patch-set: 2