Update patch set 1
Patch Set 1: Verified+1

> shouldn't this also update patches/bcprov.patch ?

The original change that this modifies isn't in the patchfile, so I suspect the file is significantly out of date (it appears to be a diff from 1.52, and we're currently on 1.56).  I've filed b/62046996 to track that problem, I'd rather not block this on getting that figured out.

Patch-set: 1
Reviewer: Gerrit User 1042885 <1042885@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1