Use generated source not jarjar
This commit switches to generating source code for
com.android.org.bouncycastle used on Android devices rather than relying
on jajar to repackage the bytecode.

Source code can (and must) be regenerated using scripts in
external/bouncycastle/srcgen when changes are made to bouncycastle
source code.

Having source code for the Android internal bouncycastle is being done
to enable the code to be included in stubs for system and other internal
API surfaces: the stubs generation relies on having source (which can be
annotated, or just have associated API metadata).

Test: build
Test: CtsLibcoreTestCases
Change-Id: I6d21774088e73c936e99059a9a0f071080174015