Make the bouncycastle build targets more regular
Before this commit:

The "bouncycastle" and "bouncycastle-bcpkix" targets do different
things on host and on target WRT to the package the compiled source ends
up in. On host the library is in org.bouncycastle while on target it is
in com.android.org.bouncycastle.

There are also "bouncycastle-unbundled" and
"bouncycastle-pcpkix-unbundled" which are in org.bouncycastle.

This commit makes "bouncycastle" and "bouncycastle-bcpkix" consistently
use com.android.org.bouncycastle. All users wanting the original package
should use "bouncycastle-unbundled" and "bouncycastle-pcpkix-unbundled"
instead.

Bug: 111734251
Bug: 111055375
Test: Build only
Change-Id: I16e3464d7d6ad0fc138353cbfa3ab572318484c1