Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > FWIW - I'm going to hold off on this for a while because it may not be needed initially. No problems in principle, though? Mostly wondering about your process for updating bouncycastle from upstream.
> 
> Looks fine in principle to me.  Updating BC shouldn't be affected beyond having to run the script after syncing the new code.
> 
> One thing that would be nice to have is a sanity test that tries to ensure the two don't get out of sync.  The easiest would be to just depend on both and assert that `new org.bouncycastle.jce.provider.BouncyCastleProvider().getVersion()` and `new com.android.org.bouncycastle.jce.provider.BouncyCastleProvider().getVersion()` are the same, which should catch the majority of cases where we fail to run the update script.

The re-gen deletes and starts again, so the way to tell if they're in sync is to re-run the generation and see what git says.

I've always thought it would be nice to have a presubmit check that does that somehow, or even simpler, that checks that every file touched in <input> is also touched in <output>. Never had the time to do it, though.

Patch-set: 6
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0