Update patch set 6
Patch Set 6:

> Sure, but until we have that, it'd be nice to have something that is easy to write and can run in the continuous builds to avoid the most obvious way the files would get out of sync.

I don't think a full upgrade that changes the version number is the most likely. You, or somebody else on the core libs team is going to handle an update and there's a process we can follow.

From ICU4J experience, I'd say the most likely is somebody editing ONLY the android_ copy to make a small fix without knowing to change the original because they don't know they have to (despite "THIS IS GENERATED CODE DO NOT EDIT" at the top and however many README.android files).

If they self approve / somebody is asleep at the wheel then that goes in. The suggested test wouldn't catch that. It *might* be caught the next time somebody runs the generation but it could be lost in the noise and then if they submit that the android_ change is lost.

Patch-set: 6
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0