Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> I presume this is necessary for our new tooling?  (I'm pretty sure you know that they were previously hidden via blanket rule rather than needing individual @hides.)

bouncycastle doesn't contribute to the public SDK so previously we just didn't include it in the set of files we use to generate that. bouncycastle will contribute to the core platform API now, though.

It's easiest if we have a single list of files for both compilation and API generation and not have separate lists of "internal, non-api files" and "api files".

As a counter example, for core-oj we're heading towards "api files", "internal apis only" and "compilation only" lists, which is a bit fiddly.

Since we generate the source for bouncycastle it made most sense to me just at @hide to everything to make it clear they're not part of the public SDK api. Some of this is shaking out now and it may go a different way, but it's easy enough to reverse later. We can start adding CorePlatformApi and IntraCoreApi as needed.

Patch-set: 2
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>