Move android_bcprov to repackaged/bcprov
This is needed to provide consistency between bouncycastle and other
users of RepackagingTransform to allow them to switch to using a common
script.

The previous change to switch from jarjar to currysrc did not correctly
update the build exclusion of the repackaged version of
org/bouncycastle/asn1/ocsp which is not required on Android. That meant
it was incorrectly being included in the bouncycastle library. This
updates the build script to remove the repackaged versions of those
files altogether and removes the unnecessary build exclusion.

Tested by running the script and verifying that apart from the now
deleted files the repackaged files only differed by location and not
content.

Bug: 117818301
Test: see above
Change-Id: I88eef5583071156c277ac82bb89ec76a91603df9