Set default visibility to private
Takes advantage of a new mechanism for specifying the default visibility
of all modules in a package to restrict the visibility of bouncycastle
modules as much as possible:
  package {default_visibility: ["//visibility:private"]}

Moved the bouncycastle-unbundled module together with the other
unbundled modules. Removed the legacy bouncycastle-bcpkix-host target as
it is not used and left the bouncycastle-host target as private
visibility so that when the existing .mk usages are switched to .bp
the references can be replaced with bouncycastle-unbundled.

Bug: 133290645
Test: m droid
Change-Id: I59e79cee159a58c5619b2e501a4bd5276756c014