Set default visibility to private
Takes advantage of a new mechanism for specifying the default visibility
of all modules in a package to restrict the visibility of bouncycastle
modules as much as possible:
  package {default_visibility: ["//visibility:private"]}

Moved the bouncycastle-unbundled module together with the other
unbundled modules. Removed the legacy bouncycastle-bcpkix-host target as
it is not used and left the bouncycastle-host target as private
visibility so that when the existing .mk usages are switched to .bp
the references can be replaced with bouncycastle-unbundled.

Added visibility to bouncycastle-testdex even though it is only used in
.mk files to record the dependency as the name is constructed
dynamically by adding "-testdex" suffix to a list of other dependencies
and so cannot be found by searching.

Bug: 133290645
Bug: 134379140
Test: m droid
Change-Id: I59e79cee159a58c5619b2e501a4bd5276756c014