Update patch set 4
Patch Set 4:

PTAL. I found that bouncycastle-testdex is used by art tests but not explicitly mentioned in either a .bp or .mk file. I though it was worth adding a visibility entry to record that usage.

Converting the .mk files to use .bp files is going to be a nightmare.

Patch-set: 4
Reviewer: Gerrit User 1039338 <1039338@85c56323-6fa9-3386-8a01-6480fb634889>