Cherrypick BigIntegers code from BouncyCastle v1.62.
Contains performance fixes to this class made between
BC version 1.61 (latest in AOSP) and 1.62 (latest public release)
in response to https://github.com/bcgit/bc-java/issues/484

Combines 3 BC commits which together apply cleanly to AOSP:-
https://github.com/bcgit/bc-java/commit/f72dd46
https://github.com/bcgit/bc-java/commit/522f6de
https://github.com/bcgit/bc-java/commit/a34ea0e

Does not contain the regression test added as part of those commits
as AOSP does not use or contain the relevant test framework.

Patch applies cleanly, and so should not cause any problems
when AOSP is updated to the rest of BC 1.62, or later.

Speedup on Taimen seems to be about 3x, see bug for detailed numbers.

Bug: 136447095
Test: atest CtsLibcoreTestCases
Change-Id: Iacd883187cc8249d287c8e5ef92ce7811d4cd540