Update patch set 1
Patch Set 1:

(1 comment)

Feels like this is a good performance fix so I approve on general principle. I've not tried to understand the actual change. Shout if you want me to look closer.

We can chat offline if you like about the impact of the performance regression/fix, any associated risk and therefore whether it should be cherry-picked anywhere.

Are you planning any follow up changes / process steps so we spot similar regressions in future? Seems like we found this one because I happened to notice / mention it and concocted the graphs to prove something weird was happening and you were able to correlate it with the bouncycastle upgrade.

Even when running well this test takes a substantial proportion of CtsLibcoreTestCases so I think it could be split out for the benefit of several folks.

Patch-set: 1
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>