Add new bouncycastle java library for use by IKE.
IKEv2 and EAP require the use of SHA1 and MD4 hash algorithms to be
standards compliant when supporting  legacy algorithms. Specifically,
EAP-SIM and EAP-AKA require SHA1 for implementing the FIPS 186-2
algorithm in order to compute the master session key. EAP-MSCHAPv2
requires MD4 for hashing the user's password and hashing the
password-hash.

Instead of using the existing bouncycastle-unbundled library, a new
library will be defined for use only by the IKE project. This ensures
that the jar size is as small as possible, as it must be loaded into the
bootpath a long with IKE.

Bug: 142345379
Test: compiles with aosp/1182882
Change-Id: Ib89d21a27f527c731eb9ddaa153705aaf7093f1d