Revert "Extend the visibility of bouncycastle to all of libcore"
Revert "Clarify where GPL-v2 and GPL-v2-with-classpath-exception..."

Revert "Allow all of libcore to have NewApi warnings"

Revert "Extend the visibility of okhttp to all of libcore"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/44d0f8f8-2169-4cd8-853a-8510415cd49e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7c0f2739376feb8e43931011240da7c8784cafd6