Revert "Extend the visibility of bouncycastle to all of libcore"
Revert "Clarify where GPL-v2 and GPL-v2-with-classpath-exception..."

Revert "Allow all of libcore to have NewApi warnings"

Revert "Extend the visibility of okhttp to all of libcore"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f3099fc5-b626-4492-a346-243a039ef0e9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I57571578b96912086d0ab67f766a6ded0ddc9e81