Revert "Extend the visibility of bouncycastle to all of libcore"
Revert "Clarify where GPL-v2 and GPL-v2-with-classpath-exception..."

Revert "Allow all of libcore to have NewApi warnings"

Revert "Extend the visibility of okhttp to all of libcore"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/c835dbe9-73e2-4f94-a190-cd21b5bb08c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6d5ba4413627472340e043e9986a48d3aa9f3ec0