Revert "Extend the visibility of bouncycastle to all of libcore"
Revert "Clarify where GPL-v2 and GPL-v2-with-classpath-exception..."

Revert "Allow all of libcore to have NewApi warnings"

Revert "Extend the visibility of okhttp to all of libcore"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/a860da0d-232d-4914-adac-cef7de116562). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9ab8b84973b63fb0d58a22c06d67c4dd3ed0dfa7