Revert "Don't throw exceptions from BCPrivate provider."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bf5083a1-f767-47d9-9f1c-5df1b19f0143). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib4dc9591fc0c6d9797e12502a4cdecafce4ad393