Revert "Reject BC keys without IV."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5bec1262-e636-4c49-9624-0316c00c1ad7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88230ea18a0bddd0c8913900ec0adf98c7854cc6