Revert "Reject BC keys without IV."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e250c326-3609-4d72-9e10-153eeaa5886d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If41dc7eb86059141d814bbaca920d4b75be9ed87