Revert "Run R8 on bouncycastle with shrinking for the boot class..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b3bb3785-a864-44ca-bd9b-ecfb915e0ce8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9843bef823ffe8d0365fbcbf989bd1080c5d8604