Revert "Disable r8 minimisation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f4cf1d67-e12d-4074-a60d-60486a7f06c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieb7536f76dbe9ff08fe5cca704244ce6fd19fe66
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com